# lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SWARM-FORGE Control Plane API
 * OpenAPI spec version: 0.1.0
 */

export interface ApiError {
  error: string;
  message: string;
}