# lib/api-zod/src/generated/types/createSwarmRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SWARM-FORGE Control Plane API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSwarmRequestModelTier } from "./createSwarmRequestModelTier";

export interface CreateSwarmRequest {
  /**
   * Natural language task description
   * @minLength 10
   * @maxLength 2000
   */
  task: string;
  /**
   * @minimum 1
   * @maximum 16
   */
  maxWorkers?: number;
  modelTier?: CreateSwarmRequestModelTier;
  /** Use mock/demo mode (no API key required) */
  mockMode?: boolean;
}