# lib/api-zod/src/generated/types/firewallVerdict.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SWARM-FORGE Control Plane API
 * OpenAPI spec version: 0.1.0
 */

export interface FirewallVerdict {
  passed: boolean;
  blockedPattern?: string | null;
  droppedCapabilities: string[];
  unicodeDetected: boolean;
  base64EntropyScore: number;
  overridePatternFound: boolean;
  verdictMessage: string;
}