# lib/api-zod/src/generated/types/swarm.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SWARM-FORGE Control Plane API
 * OpenAPI spec version: 0.1.0
 */
import type { SwarmNode } from "./swarmNode";
import type { SwarmStatus } from "./swarmStatus";

export interface Swarm {
  id: string;
  task: string;
  status: SwarmStatus;
  mockMode: boolean;
  nodes: SwarmNode[];
  edges: string[][];
  criticalPath: string[];
  parallelGroups: string[][];
  nodesTotal: number;
  nodesSucceeded: number;
  nodesFailed: number;
  tokensConsumed: number;
  executionDurationMs?: number;
  createdAt: Date;
  updatedAt: Date;
  errorMessage?: string | null;
}